Fix compiler warning for 32-bit builds.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@26956 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogDialog/LogDlg.cpp
@@ -1898,7 +1898,7 @@ void CLogDlg::LogThread()
             if (m_bMonitoringMode && m_revUnread)
             {
                 // in monitoring mode, select the first _unread_ revision
-                for (; selIndex < m_logEntries.size(); ++selIndex)
+                for (; selIndex < (int)m_logEntries.size(); ++selIndex)
                 {
                     if (m_logEntries.GetVisible(selIndex)->GetRevision() < m_revUnread)
                         break;
